Add option to allow an RX to select the modes it supports
For now, an RX can choose between dynamic 1-5, 2-5 and any fixed
mode. For now, the information is only transmitted when binding so
any changes to the settings require a rebind. That will be improved
in the following commits by using RMP to relay the supported modes
to the TX when they're changing during live operation.

Also, add a new air_lora_config_t type which encapsulates the long
range radio driver, the band and the supported modes. This way
IO RC drivers can take just one pointer to a parameter of this type
rather than multiple ones for the long range setup.