Disable HW brownout detector
When using a TX setup with a diode, TTGO v2 modules might experience
a very short drop in voltage on startup due to the OLED not having an
RST pin and starting up with all pixels in white at the same time
as the other subsystem. Disabling the brownout detector allows the
module to boot up and it doesn't cause any problems because the drop
is very short lived.