Implement frequency error correction in the RX side
Thanks to hdphillip for the idea and all tests showing that
this was worth implementing.

This ensures that both TX and RX will converge to exactly the same
frequency, regardless of the xtal frequency differences, also
accounting for temperature induced drift.

Since the frequency error has to be sampled after the TX packet
has been received, this increases the processing time by a few us,
but it's enough to make MODE 4 not work at 55ms anymore, so it
has been increased to a 57ms period.