Add support for disabling GPIO remapping at compile time
For some targets it doesn't make much sense, since the pins they
hook up to are fixed (e.g. TX). Also, some platforms won't support
it at all (e.g. UART remapping is not possible on STM32). This
also saves a bit of flash and ram on platforms that don't need
this functionality.